detect and remove spring files

+++ cg/meta/clean/demultiplexed_flow_cells.py
@@ -4,7 +4,7 @@ import logging
 import re
 import shutil
 from pathlib import Path
-from typing import Iterable, List
+from typing import Iterable, List, Optional
 
 from housekeeper.store import models as hk_models
 
@@ -32,18 +32,23 @@ class DemultiplexedRunsFlowCell:
         flow_cell_path: Path,
         status_db: Store,
         housekeeper_api: HousekeeperAPI,
+        spring_file_paths: Optional[List[str]] = None,
     ):
+        self.path: Path = flow_cell_path
         self.status_db: Store = status_db
         self.hk: HousekeeperAPI = housekeeper_api
-        self.path: Path = flow_cell_path
+        self.spring_file_paths: Optional[List[str]] = spring_file_paths
         self.run_name: str = self.path.name
         self.split_name: List[str] = re.split("[_.]", self.run_name)
         self.identifier: str = self.split_name[FLOW_CELL_IDENTIFIER_POSITION]
         self.id: str = self.identifier[1:]
         self._hk_fastq_files = None
+        self._hk_spring_files = None
         self._is_correctly_named = None
         self._exists_in_statusdb = None
         self._fastq_files_exist_in_housekeeper = None
+        self._spring_files_exist_in_housekeeper = None
+        self._files_exist_in_housekeeper = None
         self._fastq_files_exist_on_disk = None
         self._passed_check = None
 
@@ -55,6 +60,16 @@ class DemultiplexedRunsFlowCell:
                 tags=[self.id, HousekeeperTags.FASTQ]
             )
         return self._hk_fastq_files
+
+    @property
+    def hk_spring_files(self) -> List[hk_models.File]:
+        """All spring files in Housekeeper for a particular flow cell"""
+        if not self._hk_spring_files:
+            self._hk_spring_files = []
+            for spring_file_path in self.spring_file_paths:
+                if self.id in spring_file_path:
+                    self._hk_spring_files.append(spring_file_path)
+        return self._hk_spring_files
 
     @property
     def is_correctly_named(self) -> bool:
@@ -85,6 +100,30 @@ class DemultiplexedRunsFlowCell:
         return self._fastq_files_exist_in_housekeeper
 
     @property
+    def spring_files_exist_in_housekeeper(self) -> bool:
+        """Checks if the flow cell has any spring files in Housekeeper"""
+        if self._spring_files_exist_in_housekeeper is None:
+            self._spring_files_exist_in_housekeeper: bool = any(
+                [self.id in path for path in self.spring_file_paths]
+            )
+            if not self._spring_files_exist_in_housekeeper:
+                LOG.warning("Flow cell %s does not have any spring files in Housekeeper!", self.id)
+        return self._spring_files_exist_in_housekeeper
+
+    @property
+    def files_exist_in_housekeeper(self) -> bool:
+        """Checks if the flow cell has any files, either fastq or spring, in Housekeeper"""
+        if self._files_exist_in_housekeeper is None:
+            self._files_exist_in_housekeeper: bool = (
+                self.fastq_files_exist_in_housekeeper or self.spring_files_exist_in_housekeeper
+            )
+            if not self._files_exist_in_housekeeper:
+                LOG.warning(
+                    "Flow cell %s has neither fastq files nor spring files in Housekeeper!", self.id
+                )
+        return self._files_exist_in_housekeeper
+
+    @property
     def fastq_files_exist_on_disk(self) -> bool:
         """Checks if the fastq files that are in Housekeeper are actually present on disk"""
         if self._fastq_files_exist_on_disk is None:
@@ -111,7 +150,7 @@ class DemultiplexedRunsFlowCell:
                 [
                     self.is_correctly_named,
                     self.exists_in_statusdb,
-                    self.fastq_files_exist_in_housekeeper,
+                    self.files_exist_in_housekeeper,
                     self.fastq_files_exist_on_disk,
                 ]
             )
@@ -125,13 +164,20 @@ class DemultiplexedRunsFlowCell:
         return self._passed_check
 
     def remove_files_from_housekeeper(self):
-        """Remove fastq files and the sample sheet from Housekeeper when deleting a flow cell from
-        demultiplexed-runs"""
-        for fastq_file in self.hk_fastq_files:
-            self.hk.delete_file(fastq_file.id)
-        sample_sheets = self.hk.files(tags=[self.id, HousekeeperTags.SAMPLESHEET])
-        for sample_sheet in sample_sheets:
-            self.hk.delete_file(sample_sheet.id)
+        """Remove spring files, fastq files and the sample sheet from Housekeeper when deleting a
+        flow cell from demultiplexed-runs"""
+        if self.fastq_files_exist_in_housekeeper:
+            for fastq_file in self.hk_fastq_files:
+                LOG.info(f"Removing {fastq_file} from Housekeeper.")
+                self.hk.delete_file(fastq_file.id)
+            sample_sheets = self.hk.files(tags=[self.id, HousekeeperTags.SAMPLESHEET])
+            for sample_sheet in sample_sheets:
+                LOG.info(f"Removing {sample_sheet} from Housekeeper.")
+                # self.hk.delete_file(sample_sheet.id)
+        if self.spring_files_exist_in_housekeeper:
+            for spring_file in self.hk_spring_files:
+                LOG.info(f"Removing {spring_file} from Housekeeper.")
+                self.hk.delete_file(spring_file.id)
 
     def remove_from_demultiplexed_runs(self):
         """Removes a flow cell directory completely from demultiplexed-runs"""
@@ -142,5 +188,5 @@ class DemultiplexedRunsFlowCell:
     def remove_failed_flow_cell(self):
         """Performs the two removal actions for failed flow cells"""
         self.remove_from_demultiplexed_runs()
-        if self.fastq_files_exist_in_housekeeper and self.is_correctly_named:
+        if self.files_exist_in_housekeeper and self.is_correctly_named:
             self.remove_files_from_housekeeper()
